#58: pkg renamed
