+++ src/test/java/org/eolang/compiler/package-info.java
@@ -21,35 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang;
-
-import java.io.PrintStream;
-import org.apache.commons.io.output.ByteArrayOutputStream;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.Test;
 
 /**
- * Test case for {@link Main}.
+ * EO Compiler, tests.
+ *
  * @author Yegor Bugayenko (yegor256@gmail.com)
  * @version $Id$
  * @since 0.1
  */
-public final class MainTest {
-
-    /**
-     * Main can print a simple text.
-     * @throws Exception If some problem inside
-     */
-    @Test
-    public void printsSimpleResponse() throws Exception {
-        try (final ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
-            new Main(new PrintStream(baos), "--help").exec();
-            MatcherAssert.assertThat(
-                new String(baos.toByteArray()),
-                Matchers.containsString("skeleton")
-            );
-        }
-    }
-
-}
+package org.eolang.compiler;
